requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Cors
python-dotenv
psycopg2-binary
Werkzeug
gunicorn
openai